remove: info 패키지 생성 #21
