+++ src/main/java/com/windfall/api/auction/dto/response/info/scheduledInfo.java
@@ -1,10 +1,10 @@
-package com.windfall.api.auction.dto.response;
+package com.windfall.api.auction.dto.response.info;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.time.LocalDateTime;
 
-@Schema(description = "경매 진행 중 응답 DTO")
-public record processInfo(
+@Schema(description = "경매 예정 응답 DTO")
+public record scheduledInfo(
     @Schema(description = "경매 ID")
     Long auctionId,
 
@@ -17,16 +17,11 @@ public record processInfo(
     @Schema(description = "시작가")
     int startPrice,
 
-    @Schema(description = "현재가")
-    int currentPrice,
-
-    @Schema(description = "하락 퍼센트")
-    double discountRate,
-
     @Schema(description = "사용자 찜 여부")
     boolean isLiked,
 
     @Schema(description = "경매 시작 시간")
     LocalDateTime startedAt
 ) {
+
 }
